BACKUP/autonomous_agent/requirements.txt
requests>=2.28.0
pydantic>=2.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
